aprendendo mais jwt guards e tokens

+++ src/tasks/tasks.service.ts
@@ -1,9 +1,15 @@
 /*eslint-disable*/
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  HttpException,
+  HttpStatus,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { PaginationDto } from 'src/common/dto/pagination.dto';
+import { PayloadTokenDto } from 'src/auth/dto/payload-token.dto';
 
 @Injectable()
 export class TasksService {
@@ -35,14 +41,14 @@ export class TasksService {
     throw new HttpException('Tarefa não foi encontrada!', HttpStatus.NOT_FOUND);
   }
 
-  async create(createTaskDto: CreateTaskDto) {
+  async create(createTaskDto: CreateTaskDto, tokenPayload: PayloadTokenDto) {
     try {
       const newTask = await this.prisma.task.create({
         data: {
           name: createTaskDto.name,
           description: createTaskDto.description,
           completed: false,
-          userId: createTaskDto.userId,
+          userId: tokenPayload.sub,
         },
       });
 
@@ -56,12 +62,23 @@ export class TasksService {
     }
   }
 
-  async update(id: number, updateTaskDto: UpdateTaskDto) {
+  async update(
+    id: number,
+    updateTaskDto: UpdateTaskDto,
+    tokenPayload: PayloadTokenDto,
+  ) {
     try {
       const findTask = await this.prisma.task.findFirst({ where: { id: id } });
 
       if (!findTask) {
         throw new HttpException('Essa tarefa não existe', HttpStatus.NOT_FOUND);
+      }
+
+      if (findTask.userId != tokenPayload.sub) {
+        throw new HttpException(
+          'Essa tarefa não pode ser atualizada',
+          HttpStatus.UNAUTHORIZED,
+        );
       }
 
       const task = await this.prisma.task.update({
@@ -86,12 +103,19 @@ export class TasksService {
     }
   }
 
-  async delete(id: number) {
+  async delete(id: number, tokenPayload: PayloadTokenDto) {
     try {
       const findTask = await this.prisma.task.findFirst({ where: { id: id } });
 
       if (!findTask) {
         throw new HttpException('Essa tarefa não existe', HttpStatus.NOT_FOUND);
+      }
+
+      if (findTask.userId != tokenPayload.sub) {
+        throw new HttpException(
+          'Erro ao deletar essa tarefa',
+          HttpStatus.BAD_REQUEST,
+        );
       }
 
       await this.prisma.task.delete({ where: { id: findTask.id } });
