+++ src/users/users.service.ts
@@ -3,6 +3,7 @@ import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { HashingServiceProtocol } from 'src/auth/hash/hashing.service';
+import { PayloadTokenDto } from 'src/auth/dto/payload-token.dto';
 
 @Injectable()
 export class UsersService {
@@ -50,7 +51,11 @@ export class UsersService {
     }
   }
 
-  async update(id: number, updateUserDto: UpdateUserDto) {
+  async update(
+    id: number,
+    updateUserDto: UpdateUserDto,
+    tokenPayload: PayloadTokenDto,
+  ) {
     try {
       const user = await this.prisma.user.findFirst({
         where: { id: id },
@@ -58,6 +63,10 @@ export class UsersService {
 
       if (!user) {
         throw new HttpException('Usuário não existe', HttpStatus.BAD_REQUEST);
+      }
+
+      if (user.id != tokenPayload.sub) {
+        throw new HttpException('Acesso negado', HttpStatus.BAD_REQUEST);
       }
 
       const dataUser: { name?: string; passwordHash?: string } = {
@@ -92,7 +101,7 @@ export class UsersService {
     }
   }
 
-  async delete(id: number) {
+  async delete(id: number, tokenPayload: PayloadTokenDto) {
     try {
       const user = await this.prisma.user.findFirst({
         where: { id: id },
@@ -100,6 +109,10 @@ export class UsersService {
 
       if (!user) {
         throw new HttpException('Usuário não existe', HttpStatus.BAD_REQUEST);
+      }
+
+      if (user.id != tokenPayload.sub) {
+        throw new HttpException('Acesso negado', HttpStatus.UNAUTHORIZED);
       }
 
       await this.prisma.user.delete({ where: { id: id } });
